🛠 Update Error display

+++ run.sh
@@ -6,8 +6,10 @@ asm_file=$(echo ${source_file} | sed -e "s/\.pr$/\.s/")
 obj_file=${source_file%.*}
 
 cargo run -- ${source_file} -o ${asm_file}
-gcc -g -o ${obj_file} ${asm_file}
-./${obj_file}
-echo $?
-rm ${asm_file}
-rm ${obj_file}
+if [ $? -eq 0 ]; then
+    gcc -g -o ${obj_file} ${asm_file}
+    ./${obj_file}
+    echo $?
+    rm ${asm_file}
+    rm ${obj_file}
+fi
